refactor: 💡 Update index structure for sign method and proofs
