+++ test/e2e.v2.test.ts
@@ -42,7 +42,17 @@ describe("v2 E2E Test Scenarios", () => {
       if (!proof) throw new Error("No proof!");
       const msg = wrappedDocument.signature.targetHash;
       const recoverAddress = ethers.utils.verifyMessage(msg, proof.signature);
-      expect(recoverAddress.toLowerCase() === options.verificationMethod.toLowerCase()).toEqual(true);
+      expect(recoverAddress.toLowerCase()).toStrictEqual(options.verificationMethod.toLowerCase());
+    });
+    test("throws that the proof type is not supported", async () => {
+      const wrappedDocument = wrapDocument(openAttestationDatav2);
+      const options = {
+        privateKey: "0x0123456789012345678901234567890123456789012345678901234567890123",
+        verificationMethod: "0x14791697260E4c9A71f18484C9f997B308e59325",
+        type: "NotSupportType"
+      };
+      const signed = async () => await sign(wrappedDocument, options);
+      await expect(signed()).rejects.toEqual(new Error(`Proof type: ${options.type} does not exist.`));
     });
     test("should create a wrapped v2 document when issuer contains additional data", () => {
       const wrappedDocument = wrapDocument({
